exitwhen: Print a debug message if we exit before starting up
It's not an error for the exit condition to happen before nbdkit
starts up, and the filter handles this correctly.  However we should
emit at least a debug message in this case since I was surprised at
the behaviour.  I also added a comment in the code to explain the
choice.