file: trim: Don't try BLKDISCARD if earlier FALLOC_FL_PUNCH_HOLE worked
In file_zero, we test if the operations we are trying succeed and if
so jump to a single 'out:' label where we deal with the success path
out of that function.

We did not do the same thing in file_trim.  Thus in the case where
FALLOC_FL_PUNCH_HOLE succeeds, we might fall through to trying
BLKDISCARD as well.  As it happens we probably don't do this (at
least, in Linux) because we only try BLKDISCARD for block devices, and
FALLOC_FL_PUNCH_HOLE does not work on those.  But it's a good thing to
clean up this code anyway, especially if we were to add more cases in
future.

This also adds a debug message if none of the trim methods worked,
which is also analogous to what happens in the same part of file_zero.