New filter: nbdkit-count-filter: count bytes read, written etc.
This produces a summary of the number of bytes read, written, etc
through the filter.

This is mainly of use to virt-v2v where it's commonly asked how much
data was transferred over the wire or written to disk, and we don't
currently have an easy way to answer that.  By simply adding this
filter, the numbers will be known from the virt-v2v conversion log.