tests/test-floppy.sh: Simplify this test
This test created a floppy from the plugins/ directory on the fly.
Although this is not all stored in memory, the directory is still
quite large (730M on my machine) and complex.  This is a good test but
also quite time consuming and it's not really clear what we're trying
to achieve by running it every time.

Simplify by just using the plugins/floppy/ directory, but really we
ought to have a more focused test that perhaps tries all the different
file types, large files, etc.  (I added some ideas to the TODO).