full: Remove reference to equivalence of nbdkit-readonly-filter
It's not true that nbdkit-readonly-filter acts in the same way as
nbdkit-full-plugin.  The readonly filter will always return EROFS
(mapped to NBD_EPERM on the wire), whereas this plugin will always
return ENOSPC (NBD_ENOSPC).  So remove this misleading information.

This partially reverts commit 22d6ea8940 ("New filter:
nbdkit-readonly-filter"), just the changes to this one file, minus the
addition to "SEE ALSO" section.