Introduce index-gzip filter
Implements a new filter to enable space efficient random access into
gzip-compressed files by adapting the gztool [1] application (which is itself
based on "zran.c" by zlib co-author Mark Adler).

Just like the existing "gzip.c" NBDkit filter, a single decompression pass is
required, so there is no speedup upon initialization with this new approach.

The key difference is unlike the existing gzip filter the FULL uncompressed
payload does not need to be saved to temporary file but instead the only an
index file, that could be eg, 0.3% of the size of the uncompressed archive.

This finally makes it practical to use nbdkit with very large
gzipped-compressed data, such as disk image files that are tens/hundreds of
gigabytes/terabytes created by Rescuezilla or Clonezilla without needing
temporary storage space to hold the entire uncompressed data.

Note: As of this commit, the implementation does not yet contain the code to
support the indexation itself (done by gztool right now).

[1] https://github.com/circulosmeos/gztool

test

test