common: utils: vector: Fix vector_uniq prototype and add a test
The generated <vector>_uniq function takes a comparison function.  The
parameters of the comparison function point to elements of the
underlying array <type>, and we promise not to modify these elements
in the comparison function.  Previously these parameters had type
'const <type> *p', but this changes them to '<type> const *p'.

When we create a string vector (<type> == 'char *') this avoids a
warning when passing 'char **' to the original type ('const char **')
which is not legal in C for complicated reasons best explained by
reading:
https://stackoverflow.com/questions/28783534/expected-expected-const-char-but-argument-is-of-type-char

Also add a test.

Updates: commit 9bad2c1d4485aa96761fe3bc89d48609be0e4471