common: utils: vector: Add range functions for insert, append and remove
Add new functions:

  <vector>_append_array
  <vector>_insert_array
  <vector>_remove_range

These functions let you append or insert or remove a range of values
from a vector.  We occasionally need to do this in nbdkit, but it
requires a loop.  Doing it this way requires less code and is more
efficient.

This commit does not change any code outside common/utils/ to use
these functions yet, but it does add a test.