server/main.c: Reject empty string as a plugin name or filter name
Passing literal "" to dlopen appears to open the current executable.
(This is not documented, but the behaviour seems to be the same as
when passing NULL to dlopen which is documented.  Maybe this is just a
glibc thing.)  We don't want to open the current executable as either
a plugin or a filter, so reject this.

To see this problem requires a small change to is_short_name, but the
error seen before would have been:

  $ ./nbdkit ""
  nbdkit: : /home/rjones/d/nbdkit/server/nbdkit: undefined symbol: plugin_init

or:

  $ ./nbdkit --filter="" null
  nbdkit: : /home/rjones/d/nbdkit/server/nbdkit: undefined symbol: filter_init

After this change:

  $ ./nbdkit ""
  nbdkit: plugin name cannot be an empty string
  $ ./nbdkit --filter="" null
  nbdkit: filter name cannot be an empty string