server: Use debug() instead of nbdkit_debug() consistently in the server
The debug() macro calls debug_in_server().  The difference between
debug_in_server() and nbdkit_debug() is that nbdkit_debug() messages
are highlighted in the output -- useful for a user who wants to see
messages from plugins that they wrote, amongst all the other messages
generated by the server.

Also we define nbdkit_debug as a _Pragma to force a GCC error when
used in the server, so new nbdkit_debug statements cannot be added in
future.