tests: Modify make-pki and make-psk scripts to be atomic
Modify these existing scripts so they are atomic (from the point of
view of the containing Makefile, which is the only place that they
should be run from).

They now both take a single parameter which is the temporary output
directory (or output file for PSK).  The Makefile deals with moving
the temporary to the permanent location if the script succeeds.