Restructure zran.h, zran.c for use as library
Basic restructuring to simplify using zran as a library.

Libraries often want to modify or reimplement the build and extract functions,
so this means exposing an unfortunate amount of the internal #defines etc.

This commit was originally targeted at the zlib git repository, but patch
applied to nbdkit. This change is not integrated into build script yet, but can
be built using eg, [1]

[1] gcc -DTEST -o zran zran.c -lz