tests/test-golang-fork-warning.sh: Fix hanging test
In CI, this test regularly causes the test suite to exceed the 1 hour
timeout.  Add a timeout in this test so that if it's hanging, the test
will exit after 6 minutes.

It's hard to diagnose this since (a) it's not easily reproducible and
(b) gitlab CI does not save artifacts if the test suite exceeds the
total time.  However I was able to diagnose it by diffing the list of
tests PASSed between successful and timed out runs.

We already tried to fix this in commit 8339df9db2 ("tests/
test-golang-fork-warning.sh: Don't use nbdinfo in test") but that fix
was not sufficient.

Updates: commit 8339df9db2a472e2be8170e22499ac7ec6e8272f